+++ Sources/BaseFoundationiOS/Protocols/Delegates/DataModelDelegate.swift
@@ -1,8 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by pablo borquez on 12/05/24.
-//
-
-import Foundation